requirements.txt
absl-py==1.0.0
asttokens==2.0.5
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
cachetools==5.0.0
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
contextlib2==21.6.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
execnet==1.9.0
executing==0.8.3
Flask==2.1.1
flatbuffers==2.0
fonttools==4.33.0
gast==0.5.3
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.2.2
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mock==4.0.3
nest-asyncio==1.5.5
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
path==16.4.0
path.py==12.5.0
pickleshare==0.7.5
Pillow==9.1.0
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.8
pytest==7.1.1
pytest-shutil==1.7.0
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.14.0
